Mv commentable and voteable helper mods from lib to controllers/concerns
Modify the Posts and Comments controllers as necessary to reflect the new file paths and names
