+++ app/controllers/posts_controller.rb
@@ -1,6 +1,6 @@
 class PostsController < ApplicationController
-  include HackernewsClone::VoteHelper
-  include HackernewsClone::CommentHelper
+  include VoteableController
+  include CommentableController
 
   before_action :authenticate_user!, only: :create
   before_action :authenticate_to_submit, only: :new
