--- lib/hackernews_clone/vote_helper.rb
@@ -1,11 +1,0 @@
-module HackernewsClone
-  module VoteHelper
-    def self.included(base)
-      base.helper_method :parents_vote
-    end
-
-    def parents_vote(parent)
-      @vote = Vote.find_by(user_id: current_user, voteable_id: parent)
-    end
-  end
-end
